Reacted on -Wsign-compare warnings of gcc
git-svn-id: http://svn.libburnia-project.org/libburn/trunk@3796 fff8544e-141b-0410-a922-c1ea1f00b384

+++ test/poll.c
@@ -67,7 +67,7 @@ int main()
 
 	newact.sa_handler = catch_int;
 	sigaction(SIGINT, &newact, &oldact);
-	for (i = 0; i < n_drives; i++) {
+	for (i = 0; i < (int) n_drives; i++) {
 	    NEXT=0;
 	    poll_drive(i);
 	}
